back-end/requirements.txt
django
djangorestframework
pyyaml
httpie
requests
django-cors-headers